Bug 1408100 - Upgrade to Bootstrap 4 (Trickier customizations)
These aren't all THAT tricky.  But this required more than just adding
a class that Bootstrap provided.  All the pixel-peeping to make
things look presentable.

Note: This is not pixel-perfect with Bootstrap 3.x.  I got it to where
I think it looks good, but didn't go through extreme measures to
keep everything identical.